Renamed and moved guard files
