+++ client/src/app/shared/guards/user-is-ebi.guard.ts
